+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 * Consumables can now be pulled from postmasters other than the active character's.
+* Vendors page now correctly recognizes owned bounties per character and is more accurate about mods ownership.
 
 ## 6.97.3 <span class="changelog-date">(2021-12-30)</span>